+++ packages/components/Checkbox/src/checkmark/checkmark.android.ts
@@ -1,0 +1,2 @@
+import checkmarkSvg from '../assets/checkmark_android.svg';
+export default checkmarkSvg;