+++ packages/components/ContextualMenu/src/SubmenuItem.tsx
@@ -41,8 +41,8 @@ export const SubmenuItem = compose<SubmenuItemType>({
 
     const onItemClick = React.useCallback(
       (e) => {
-        if (!disabled ) {
-          context ?.onDismissMenu();
+        if (!disabled) {
+          context?.onDismissMenu();
           onClick ? onClick() : context.onItemClick(itemKey);
           e.stopPropagation();
         }
@@ -56,9 +56,11 @@ export const SubmenuItem = compose<SubmenuItemType>({
       (e) => {
         componentRef.current.focus();
         userProps.onHoverIn(e);
-      }, [componentRef]);
+      },
+      [componentRef],
+    );
 
-    const pressable = useAsPressable({ ...rest, onPress: onItemClick, onHoverIn: onItemHoverIn});
+    const pressable = useAsPressable({ ...rest, onPress: onItemClick, onHoverIn: onItemHoverIn });
 
     // set up state
     const state: SubmenuItemState = {
@@ -69,9 +71,9 @@ export const SubmenuItem = compose<SubmenuItemType>({
       icon: !!icon,
     };
 
-      /*
-    * For SubmenuItem, menu is shown on hover.
-    */
+    /*
+     * For SubmenuItem, menu is shown on hover.
+     */
     const onKeyUp = useKeyCallback(onItemHoverIn, ' ', 'Enter', 'ArrowRight');
 
     const svgProps: SvgIconProps = {
@@ -94,7 +96,7 @@ export const SubmenuItem = compose<SubmenuItemType>({
         ...pressable.props,
         ref: cmRef,
         onKeyUp: onKeyUp,
-        accessibilityLabel: accessibilityLabel
+        accessibilityLabel: accessibilityLabel,
       },
       content: { children: text, testID },
       icon: createIconProps(icon),