+++ packages/experimental/Button/src/CompoundButton/CompoundButton.types.ts
@@ -1,4 +1,4 @@
-import { ViewProps } from 'react-native';
+import { ViewProps, ColorValue } from 'react-native';
 import { TextProps } from '@fluentui-react-native/experimental-text';
 import { ButtonSlotProps, ButtonTokens, ButtonProps } from '../Button.types';
 import { FontTokens } from '@fluentui-react-native/tokens';
@@ -7,7 +7,7 @@ export const compoundButtonName = 'CompoundButton';
 
 export interface CompoundButtonTokens extends ButtonTokens {
   secondaryContentFont?: FontTokens;
-  secondaryContentColor?: string;
+  secondaryContentColor?: ColorValue;
 }
 
 export interface CompoundButtonProps extends ButtonProps {