+++ packages/theming/win32-theme/src/NativeModule/officeThemingModule.ts
@@ -1,5 +1,5 @@
-import type { NativeModule } from 'react-native';
-import { ColorValue, OfficePalette, Typography } from '@fluentui-react-native/theme-types';
+import type { NativeModule, ColorValue } from 'react-native';
+import { OfficePalette, Typography } from '@fluentui-react-native/theme-types';
 
 export type PlatformDefaultsChangedArgs = { hostThemeSetting: string };
 export type PlatformDefaultsChangedCallback = (args?: PlatformDefaultsChangedArgs) => void;